refactor(constants): 공통으로 사용되는 상수를 하나의 클래스로 모으기
