+++ src/main/java/lotto/constants/CommonConstants.java
@@ -1,0 +1,6 @@
+package lotto.constants;
+
+public class CommonConstants {
+    public static final long UNIT_PURCHASE_AMOUNT = 1000L;
+    public static final long MAX_PURCHASE_AMOUNT = 1_000_000L;  // 실제 1인당 로또 구매가능 금액
+}
