+++ src/main/java/lotto/util/NumLottoCalculator.java
@@ -1,8 +1,8 @@
 package lotto.util;
 
-public class NumLottoCalculator {
-    private static final Long UNIT_PURCHASE_AMOUNT = 1000L;
+import static lotto.constants.CommonConstants.UNIT_PURCHASE_AMOUNT;
 
+public class NumLottoCalculator {
     public static long calculate(long purchaseAmount) {
         return purchaseAmount / UNIT_PURCHASE_AMOUNT;
     }
